+++ src/io/reader/reader_mmtf.cpp
@@ -133,7 +133,7 @@ namespace VTX
 			p_molecule.addBonds( data.bondAtomList );
 
 #ifdef _DEBUG
-			p_molecule.bondCount += data.bondAtomList.size();
+			p_molecule.bondCount += (uint)data.bondAtomList.size();
 #endif
 
 #ifdef _DEBUG
